requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.19
PyYAML==3.11
Werkzeug==0.9.4
ipython==1.2.1
itsdangerous==0.23
requests==2.2.1
wsgiref==0.1.2
